update run_cycle*pl, add RUNDIR aux file path

+++ run_cycle.plot_sfcobs_sw_txt.pl
@@ -2,7 +2,7 @@
 
 $GMID="GECN3KM";
 $MEMBER="GFS_WCTRL";
-$CYCLE="2017051012";
+$CYCLE="2017060412";
 $start_hour=0;
 $end_hour=72;
 $incre_hour=1;
@@ -23,6 +23,7 @@ require "$ENSPROCS/common_tools.pl";
 
 for ($hr=$start_hour; $hr <=$end_hour; $hr=$hr+$incre_hour) {
     $d=&tool_date12_add("${CYCLE}00", $hr, "hour");
+    print("to process $d -----\n");
     $bjd=&tool_date12_add($d, 8, "hour");
     $mywork="$WORKDIR/$CYCLE/$d/";
     system("test -d $mywork || mkdir -p $mywork");
@@ -30,6 +31,10 @@ for ($hr=$start_hour; $hr <=$end_hour; $hr=$hr+$incre_hour) {
     #cp aux
     $file_name2_nc3=&tool_date12_to_outfilename("auxhist3_d0${dom}_", "${d}00", ".nc");
     $file_temp_save="$TEMPDIR/$file_name2_nc3";
+    $file_name1=&tool_date12_to_outfilename("auxhist3_d0${dom}_", "${d}00", "");
+    $file_path1="$RUNDIR/$CYCLE/WRF_P/$file_name1";
+    $file_name2=&tool_date12_to_outfilename("auxhist3_d0${dom}_", "${d}", ".nc4.p");
+    $file_path2="$ARCDIR/aux3_$CYCLE/$file_name2";
     if( -e $file_name2_nc3) {
         print "file exists in /dev/shm\n";
         if( !-e $file_temp_save) { 
@@ -37,9 +42,9 @@ for ($hr=$start_hour; $hr <=$end_hour; $hr=$hr+$incre_hour) {
         }
     }elsif (-e $file_temp_save) { 
         system("cp -r $file_temp_save $mywork/");
-    }else {
-        $file_name2=&tool_date12_to_outfilename("auxhist3_d0${dom}_", "${d}", ".nc4.p");
-        $file_path2="$ARCDIR/aux3_$CYCLE/$file_name2";
+    }elsif (-e $file_path1) {
+        system("cp -r $file_path1 $mywork/");
+    }elsif (-e $file_path2) {
         system("cp $file_path2 $mywork/");
         $file_name2_unpack=&tool_date12_to_outfilename("auxhist3_d0${dom}_", "${d}", ".nc4");
         print("doing ncpdq unpacking..\n");
@@ -48,10 +53,17 @@ for ($hr=$start_hour; $hr <=$end_hour; $hr=$hr+$incre_hour) {
         system("ncks -O -3 $file_name2_unpack $file_name2_nc3");
         system(" test -d $TEMPDIR || mkdir -p $TEMPDIR");
         system("cp $file_name2_nc3 $TEMPDIR/");
+    }else {
+        print("aux file not found , next \n");
+        next;
     }
     $file_path="$mywork/$file_name2_nc3";
     #cp obsfile
     $obs_file="$OBSDIR/${bjd}_Radiation.txt";
+    if( ! -e $obs_file ) {
+        print("$obs_file not found, next \n");
+        next;
+    }
     system("cp $obs_file $mywork");
     print("$obs_file\n");
     #ln ncl
